terminal-config-refactor - extract config parsing into module
