+++ Makefile
@@ -50,7 +50,7 @@ all: $(binary)
 release: CPPFLAGS += -DG_DISABLE_ASSERT -DNDEBUG
 release: $(binary)
 
-$(binary): src/terminal-config.h \
+$(binary): src/terminal-config.h src/terminal-config.c \
            src/exit-status.h src/exit-status.c \
            src/stulto-terminal.h src/stulto-terminal.c \
            src/stulto-application.h src/stulto-application.c \
